Extract highlightSearchMatch to shared utility
Move the duplicated highlightSearchMatch function from TaskCard.tsx and
TaskListView.tsx into a shared searchUtils.tsx utility file. Also fixes
the regex by using a separate test regex without the global flag.
